share datasets data btwn tooltip+modal
