+++ airflow/www/static/js/dag.js
@@ -21,7 +21,7 @@
 
 import { getMetaValue } from './utils';
 import { approxTimeFromNow, formatDateTime } from './datetime_utils';
-import { openDatasetModal } from './openDatasetModal';
+import { openDatasetModal, getDatasetTooltipInfo } from './datasetUtils';
 
 function updateQueryStringParameter(uri, key, value) {
   const re = new RegExp(`([?&])${key}=.*?(&|$)`, 'i');
@@ -32,12 +32,6 @@ function updateQueryStringParameter(uri, key, value) {
 
   return `${uri}${separator}${key}=${value}`;
 }
-
-// Pills highlighting
-$(window).on('load', function onLoad() {
-  $(`a[href*="${this.location.pathname}"]`).parent().addClass('active');
-  $('.never_active').removeClass('active');
-});
 
 const dagId = getMetaValue('dag_id');
 export const dagTZ = getMetaValue('dag_timezone');
@@ -58,6 +52,21 @@ let mapIndex;
 let mapStates = [];
 let extraLinks;
 const showExternalLogRedirect = getMetaValue('show_external_log_redirect') === 'True';
+let nextDatasets = [];
+let nextDatasetsError;
+
+const setNextDatasets = (datasets, error) => {
+  nextDatasets = datasets;
+  nextDatasetsError = error;
+};
+
+// Pills highlighting
+$(window).on('load', function onLoad() {
+  $(`a[href*="${this.location.pathname}"]`).parent().addClass('active');
+  $('.never_active').removeClass('active');
+  const run = $('#next-dataset-tooltip');
+  getDatasetTooltipInfo(dagId, run, setNextDatasets);
+});
 
 const buttons = Array.from(document.querySelectorAll('a[id^="btn_"][data-base-url]')).reduce((obj, elm) => {
   obj[elm.id.replace('btn_', '')] = elm;
@@ -403,5 +412,5 @@ $('#next-run').on('mouseover', () => {
 
 $('.next-dataset-triggered').on('click', (e) => {
   const summary = $(e.target).data('summary');
-  openDatasetModal(dagId, summary || '');
+  openDatasetModal(dagId, summary, nextDatasets, nextDatasetsError);
 });
